refactors code and improves file structure so both forms can use same inputs
