--- src/pages/LoginPage/LoginPage.module.css
@@ -1,3 +1,0 @@
-.form-conatainer {
-    width: 350px;
-}
